dns: split dns.hpp into several headers (#1241)
